BugFix: 경로 문제 해결 - 카테고리 폴더 구조 변경 - 폴더 구조 변경으로 발생한, js파일에서 import 경로 문제 해결
