+++ src/styles/style.css
@@ -9,7 +9,7 @@
 
 @import url(/src/pages/BoardPage/board.css);
 @import url(/src/pages/StartPage/start.css);
-@import url(/src/pages/CategoryPage/category.css);
+@import url(/src/pages/LoginPage/category/category.css);
 
 
 /* ---------------------------------------------------------- */
